Reuse client objects in ITs
