+++ Stormpath.SDK/Stormpath.SDK.Tests.Integration/Stormpath.SDK.Tests.Integration.csproj
@@ -70,7 +70,7 @@
     <Compile Include="Async\CustomData_tests.cs" />
     <Compile Include="Async\Group_tests.cs" />
     <Compile Include="Helpers\Assertly.cs" />
-    <Compile Include="Helpers\IntegrationTestClients.cs" />
+    <Compile Include="IntegrationTestClients.cs" />
     <Compile Include="Helpers\RandomPassword.cs" />
     <Compile Include="Helpers\RandomEmail.cs" />
     <Compile Include="Helpers\RandomPassword_tests.cs" />
